Refactor Tabel block to use block.json metadata
