+++ packages/block-library/src/table/block.json
@@ -1,0 +1,88 @@
+{
+	"name": "core/table",
+	"category": "formatting",
+	"attributes": {
+		"hasFixedLayout": {
+			"type": "boolean",
+			"default": false
+		},
+		"backgroundColor": {
+			"type": "string"
+		},
+		"head": {
+			"type": "array",
+			"default": [],
+			"source": "query",
+			"selector": "thead tr",
+			"query": {
+				"cells": {
+					"type": "array",
+					"default": [],
+					"source": "query",
+					"selector": "td,th",
+					"query": {
+						"content": {
+							"type": "string",
+							"source": "html"
+						},
+						"tag": {
+							"type": "string",
+							"default": "td",
+							"source": "tag"
+						}
+					}
+				}
+			}
+		},
+		"body": {
+			"type": "array",
+			"default": [],
+			"source": "query",
+			"selector": "tbody tr",
+			"query": {
+				"cells": {
+					"type": "array",
+					"default": [],
+					"source": "query",
+					"selector": "td,th",
+					"query": {
+						"content": {
+							"type": "string",
+							"source": "html"
+						},
+						"tag": {
+							"type": "string",
+							"default": "td",
+							"source": "tag"
+						}
+					}
+				}
+			}
+		},
+		"foot": {
+			"type": "array",
+			"default": [],
+			"source": "query",
+			"selector": "tfoot tr",
+			"query": {
+				"cells": {
+					"type": "array",
+					"default": [],
+					"source": "query",
+					"selector": "td,th",
+					"query": {
+						"content": {
+							"type": "string",
+							"source": "html"
+						},
+						"tag": {
+							"type": "string",
+							"default": "td",
+							"source": "tag"
+						}
+					}
+				}
+			}
+		}
+	}
+}
