+++ packages/block-library/src/table/save.js
@@ -1,0 +1,63 @@
+/**
+ * External dependencies
+ */
+import classnames from 'classnames';
+
+/**
+ * WordPress dependencies
+ */
+import { RichText, getColorClassName } from '@wordpress/block-editor';
+
+export default function save( { attributes } ) {
+	const {
+		hasFixedLayout,
+		head,
+		body,
+		foot,
+		backgroundColor,
+	} = attributes;
+	const isEmpty = ! head.length && ! body.length && ! foot.length;
+
+	if ( isEmpty ) {
+		return null;
+	}
+
+	const backgroundClass = getColorClassName( 'background-color', backgroundColor );
+
+	const classes = classnames( backgroundClass, {
+		'has-fixed-layout': hasFixedLayout,
+		'has-background': !! backgroundClass,
+	} );
+
+	const Section = ( { type, rows } ) => {
+		if ( ! rows.length ) {
+			return null;
+		}
+
+		const Tag = `t${ type }`;
+
+		return (
+			<Tag>
+				{ rows.map( ( { cells }, rowIndex ) => (
+					<tr key={ rowIndex }>
+						{ cells.map( ( { content, tag }, cellIndex ) =>
+							<RichText.Content
+								tagName={ tag }
+								value={ content }
+								key={ cellIndex }
+							/>
+						) }
+					</tr>
+				) ) }
+			</Tag>
+		);
+	};
+
+	return (
+		<table className={ classes }>
+			<Section type="head" rows={ head } />
+			<Section type="body" rows={ body } />
+			<Section type="foot" rows={ foot } />
+		</table>
+	);
+}
