+++ packages/block-library/src/table/transforms.js
@@ -1,0 +1,51 @@
+/**
+ * WordPress dependencies
+ */
+import { getPhrasingContentSchema } from '@wordpress/blocks';
+
+const tableContentPasteSchema = {
+	tr: {
+		allowEmpty: true,
+		children: {
+			th: {
+				allowEmpty: true,
+				children: getPhrasingContentSchema(),
+			},
+			td: {
+				allowEmpty: true,
+				children: getPhrasingContentSchema(),
+			},
+		},
+	},
+};
+
+const tablePasteSchema = {
+	table: {
+		children: {
+			thead: {
+				allowEmpty: true,
+				children: tableContentPasteSchema,
+			},
+			tfoot: {
+				allowEmpty: true,
+				children: tableContentPasteSchema,
+			},
+			tbody: {
+				allowEmpty: true,
+				children: tableContentPasteSchema,
+			},
+		},
+	},
+};
+
+const transforms = {
+	from: [
+		{
+			type: 'raw',
+			selector: 'table',
+			schema: tablePasteSchema,
+		},
+	],
+};
+
+export default transforms;
